SRS2 Use Cases Updated
Added use cases that were not included, changed the use cases to use the same format, changed wording on some cases to have more consistency. Main problem now is the numbering indentation between cases does not align. This would take a fair amount of effort to fix with little gain, so I won't do that for now, if you guys think it is important though I will in the future.