Unofficial Progress Report
I mostly wrote this for myself to better understand where we are at but I thought I should share it in case it might help someone. Note that this is a result me quickly going over the code. When I say something is implemented I mean it looks like it is implemented I have not tested it. I also wrote in some questions I thought of while reading. Feel free to ignore or delete if you think it is too much clutter.